Updates scripts

+++ 02_work_queues/worker.php
@@ -10,7 +10,6 @@ $channel->queue_declare('task_queue', false, true, false, false);
 
 echo " [*] Waiting for messages. To exit press CTRL+C\n";
 
-
 $callback = function ($msg) {
     echo ' [x] Received ', $msg->body, "\n";
     $sleepTime = substr_count($msg->body, '.');
@@ -20,7 +19,7 @@ $callback = function ($msg) {
     $msg->ack();
 };
 
-$channel->basic_qos(null, 1, null);
+$channel->basic_qos((int)null, 1, null);
 $channel->basic_consume('task_queue', '', false, false, false, false, $callback);
 
 try {
